Update openssl-ibmca to version 2.4.1 / rev 51 via SR 1187530
https://build.opensuse.org/request/show/1187530
by user ngueorguiev + anag+factory

+++ packages/o/openssl-ibmca/openssl-ibmca.spec
@@ -162,9 +162,11 @@ cp -p %{_datadir}/%{name}/openssl-ibmca.sectiondef.txt %{sslengcnf}/openssl-ibmc
 cp -p %{_datadir}/%{name}/openssl-ibmca.enginedef.cnf %{sslengdef}/openssl-ibmca.cnf
 
 %if "%{flavor}" == ""
+if [ -f "/usr/share/doc/packages/openssl-ibmca/ibmca-engine-opensslconfig" ]; then
    cp -p /usr/share/doc/packages/openssl-ibmca/ibmca-engine-opensslconfig /usr/share/doc/packages/openssl-ibmca/ibmca-engine-opensslconfig.orig
    sed -e 's/ossl-modules/engines-3/' /usr/share/doc/packages/openssl-ibmca/ibmca-engine-opensslconfig.orig > /usr/share/doc/packages/openssl-ibmca/ibmca-engine-opensslconfig
    rm /usr/share/doc/packages/openssl-ibmca/ibmca-engine-opensslconfig.orig
+fi
 %endif
 
 %postun
